reorganize code to separate concerns more clearly
